Move ServiceError regression test to AcsWebhooks namespace, resolve merge conflict
