+++ Adyen/AcsWebhooks/Models/ServiceError.cs
@@ -37,6 +37,7 @@ namespace Adyen.AcsWebhooks.Models
         /// <summary>
         /// Initializes a new instance of the <see cref="ServiceError" /> class.
         /// </summary>
+
         public ServiceError()
         {
             OnCreated();
